Move sdk file into lib folder
